requirements.txt
python-telegram-bot==20.7
rapidfuzz>=3.14
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
asyncpg